requirements.txt
fastapi
uvicorn
requests
python-telegram-bot==22.0